Restructure the vertx and client (HttpClient) usage
Initialize vertx and client (HttpClient) at the top of the class rather than before the test suite class is run.

The client and vertx shouldn't be stored in TestBase if it is already statically allocated.
Just call the method via the class (the getter name is short enough).
This reduces the code complexity.

This begins a series of changes regarding avoiding the circular logic with calling the BeforeClass and AfterClass Junit annotations for StorageTestSuite class.
The invokeStorageTestSuiteAfter boolean is removed.
This reduces the complexicity of the tests, increases the maintainability of the tests, and hopefully reduces the unpredictability of the test failures.